@@ -12,7 +12,7 @@
     }
   ],
   "require": {
-    "php": ">=7.3",
+    "php": ">=8.0.2",
     "ext-json": "*",
     "symfony/validator": "^3 || ^4 || ^5 || ^6"
   },
